# Fix the O(n^3) performance issue in `is_connected` utility function