--- a/include/boost/graph/graph_utility.hpp
+++ b/include/boost/graph/graph_utility.hpp
@@ -14,7 +14,6 @@
 #include <stdlib.h>
 #include <iostream>
 #include <algorithm>
-#include <vector>
 #include <assert.h>
 #include <boost/config.hpp>
 #include <boost/tuple/tuple.hpp>
@@ -22,9 +21,10 @@
 #include <boost/graph/graph_traits.hpp>
 #include <boost/graph/iteration_macros.hpp>
 #include <boost/graph/properties.hpp>
-#include <boost/graph/strong_components.hpp>
 #include <boost/pending/container_traits.hpp>
 #include <boost/graph/depth_first_search.hpp>
+// only for backward-compatibility error message
+#include <boost/graph/is_connected.hpp>
 // iota moved to detail/algorithm.hpp
 #include <boost/detail/algorithm.hpp>
 
@@ -341,55 +341,7 @@ inline bool is_reachable(
     return get(color, y) != color_traits< ColorValue >::white();
 }
 
-template < typename Graph, typename VertexColorMap >
-inline bool is_connected_dispatch(const Graph& g, VertexColorMap color,
-    undirected_tag)
-{
-    typedef typename property_traits< VertexColorMap >::value_type ColorValue;
-    typedef color_traits< ColorValue > Color;
-    typename graph_traits< Graph >::vertex_iterator ui, ui_end, ci, ci_end;
-
-    // clear the colormap to preserve backward compatibility
-    for (boost::tie(ci, ci_end) = vertices(g); ci != ci_end; ++ci)
-        put(color, *ci, Color::white());
-
-    default_dfs_visitor vis;
-    detail::depth_first_visit_impl(g, detail::get_default_starting_vertex(g),
-        vis, color, detail::nontruth2());
-
-    // If an undirected graph is connected, then each vertex is reachable in a
-    // single DFS visit. If any vertex was unreachable, grpah is not connected.
-    for (boost::tie(ui, ui_end) = vertices(g); ui != ui_end; ++ui)
-        if (get(color, *ui) == Color::white())
-            return false;
-    return true;
-}
-
-template < typename Graph, typename VertexColorMap >
-inline bool is_connected_dispatch(const Graph& g, VertexColorMap, directed_tag)
-{
-    // Run the Tarjan SCC algorithm
-    std::vector< size_t > comp_map(num_vertices(g));
-    strong_components(
-        g, make_iterator_property_map(comp_map.begin(), get(vertex_index, g)));
-
-    // If the directed graph is strongly connected, all vertices are in
-    // the same component 0
-    for (std::vector< size_t >::const_iterator i = comp_map.begin();
-         i != comp_map.end(); ++i)
-        if (*i > 0)
-            return false;
-    return true;
-}
-
-// Is the undirected graph connected?
-// Is the directed graph strongly connected?
-template < typename VertexListGraph, typename VertexColorMap >
-inline bool is_connected(const VertexListGraph& g, VertexColorMap color)
-{
-    typedef typename graph_traits< VertexListGraph >::directed_category Cat;
-    return is_connected_dispatch(g, color, Cat());
-}
+// is_connected is moved to `is_connected.hpp`
 
 template < typename Graph >
 bool is_self_loop(