@Wavetrace 2024-11-12
I've also opened https://github.com/boostorg/graph/issues/398 to track this.