@Wavetrace 2024-11-14
For the record, this function was added in https://github.com/boostorg/graph/commit/86ef70716434ea1dabe2b6a25bb489e19793ebfe on `Feb 6, 2001` and not modified since then except for style.