@jeremy-murphy 2024-12-11
Apologies for the delay in replying, I'm in the midst of relocating to a different country.

> > It perhaps has to be a parameter to is_connected so that the user can specify it at each call. What do you think of this situation? Let's resolve this design question before we move on to matters of code.
> 
> Yes, I like this idea. Let me draft an implementation do discuss in greater details. I'd still consider it important to keep the interface backward-compatible. What do you think about it?

I know backward-compatibility is often sought after or required, but I think it's given too much priority without consideration of the value provided. In this case, as I mentioned:

1. the function is undocumented
2. the interface is ambiguous

so in my opinion, backward-compatibility is not a priority.

I think it's better to draft an interface first before moving on to code, but I'm too late with that advice.
How I imagine it working is, given `is_connected<X>(Y)`, where `X` is a type and `Y` is a concept:

    X              Y              Result
                   undirected     OK
                   directed       invalid
    weakly         directed       OK
    unilaterally   directed       OK
    strongly       directed       OK

The relationship between the kinds of directed connectivity is that strongly ⇒ weakly and strongly ⇒ unilaterally, but nothing else (i.e. unilaterally does not ⇒ weakly nor vice versa). So I wouldn't use the struct/enum design you currently have, I would do

```
struct weakly {};
struct unilaterally {};
struct strongly : weakly, unilaterally {};
```
I believe that's how the graph concepts are done too, have a look around in `graph_traits.hpp` maybe. 

> Also is it possible to mark a function as deprecated in Boost? I only see the deprecated headers or comments for function deprecation.

https://en.cppreference.com/w/cpp/language/attributes/deprecated