@jeremy-murphy 2024-12-11
Hmmm, my idea of using the template parameter is not essential, btw. Using the second parameter for a tag like you have done is probably preferable and I think follows existing conventions. One benefit of the tag dispatch is that `is_connected(g)` can work for an undirected graph.