@Wavetrace 2024-12-12
This fixes #398